+++ Changelog.md
@@ -16,6 +16,7 @@ OpenCore Changelog
 - Added `SetupDelay` to configure audio setup delay
 - Reworked LogoutHook.command to support older macOS
 - Improved Duet USB legacy management for supported controllers
+- Implemented MP3 audio decoding for audio assistant support
 
 #### v0.6.4
 - Added `BlacklistAppleUpdate` to fix macOS 11 broken update optout